Split up intersection_shaper test.
